requirements.txt
Django==4.0
django-filter==23.5
djangorestframework==3.14.0
mysqlclient==2.2.4